REFACT: Introduce new process runner and move code
