+++ src/darsia/gui/ui/analysis.py
@@ -67,7 +67,7 @@ class AnalysisTab:
     def on_abort_clicked(self):
         """Handle abort button click."""
         if self.process is not None:
-            self.main_window.abort_workflow_process(self.process)
+            self.main_window.process_runner.abort_workflow_process(self.process)
 
     def run_analysis(self):
         """Run analysis workflow based on checked checkboxes."""
@@ -125,6 +125,6 @@ class AnalysisTab:
             argv.append("--show")
 
         # Launch workflow in a separate process
-        self.process = self.main_window.start_workflow_process(
+        self.process = self.main_window.process_runner.start_workflow_process(
             argv, self.run_button, self.abort_button, cwd=Path.cwd()
         )
