+++ src/darsia/gui/ui/setup.py
@@ -64,7 +64,7 @@ class SetupTab:
     def on_abort_clicked(self):
         """Handle abort button click."""
         if self.process is not None:
-            self.main_window.abort_workflow_process(self.process)
+            self.main_window.process_runner.abort_workflow_process(self.process)
 
     def run_setup(self):
         """Run setup workflow based on checked checkboxes."""
@@ -168,6 +168,6 @@ class SetupTab:
             argv.append("--show")
 
         # Launch workflow in a separate process
-        self.process = self.main_window.start_workflow_process(
+        self.process = self.main_window.process_runner.start_workflow_process(
             argv, self.run_button, self.abort_button, cwd=Path.cwd()
         )
